fix the ordering of same units with different powers
When working with units internally we can have the same units with
different powers in a `CTCompoundUnit`. And for certain units,
e.g. Radian or Steradian we need to keep this information alive until
the end.

For these cases the ordering of the units with powers becomes
important. While we wish to respect the ordering of the unit kinds via
`<`, the powers need to be the inverse of that. A larger power
is *smaller* (appears to the *left*) of a smaller power. Essentially
to allow intuitive ordering of fractions.

This is important when dealing with radians, because otherwise doing
math `+`, `-` with radian (as it's currently defined as `Radian =
Meter•Meter⁻¹` will emit `Meter⁻¹•Meter`.