add SI prefix factor for compound units to fix compound math
Math of compound units was broken, if it involved different SI
prefixes, as the prefix was dropped when converting to a flat
representation. E.g. `GigaTesla` would drop `Giga` if it involved math
with a different unit that converted Tesla to its base representation.

I hope this is surgical enough and doesn't leave cases that are still
broken though :/