clean up `defUnit`, also define base unit aliases if valid
This extends defUnit to also define the base type aliases of the given
unit, if valid. I.e. `J•s` also defines the equivalent `kg•m²•s⁻¹`,
but `mJ•s` does _not_ define any base unit alias. In the former a
direct type only equivalent exists, but in the latter it doesn't (as a
scaling factor is required that cannot be cleanly mapped to the type
system, i.e. where does the `milli` go that is part of the Joule in
the base units?)