change default generated definitions of units
Instead of generating the quantity types first and then all units, we
now generate all units first as based on distinct quantities. Then for
compound units we define all versions of base types as aliases to the
unit name. That way we give the compound unit name "priority" in the
eyes of the Nim compiler.