make `$`, `pretty` macros to avoid Nim compiler eating alias names
The issue with the recent changes in `defUnit` to add more aliases for
a better user experience is that generics tend to eat aliases. In a
generic setting the Nim compiler will often only give us the first non
aliased name. This is pretty annoying, as it then often gives us type
names different from what the user intended. This gets around it, as
we now perform the logic in place.