also turn `to` into a macro for the same reason as `$`
Also here the issue is the same. As a generic we lose access to the
real type of the symbol :(.

Note that this may make `toDef` partially unnecessary. We'll see how
this works out.

Also note that this breaks `to` in context of locally defined units
combined with generics, as the macro for `defUnit` won't be evaluated
before and thus `to` won't know the type.

However, this is actually a lesser problem now, because we generate
more aliases so the *need* for `to` has drastically gone down.