give each Quantity unique ID, add QuantityPowerArray for optimization
QuantityPowerArray is a helper type to simplify the code dealing with
dimensional analysis of quantities. Instead of having a "runtime"
sequence of quantities and their powers (and having to do complicated
comparisons) we simply have a "fixed size" sequence, where each index
corresponds to the i-th quantity. Allows for much simpler handling of
quantities and their powers, without the need to compare quantities
itself or sort the sequences.