rewrite `simplify` for much better performance
Instead of using `CountTables`, which are pretty slow for this
purpose, just work with the sorted units directly and accumulating all
powers of the same unit (either ignoring or not the SI prefix).

NOTE: this can arguably be improved further by an approach similar to
QuantityPowerArray so that we get accumulation automatically by
inserting each unit into its appropriate field!