fix regression due to another fallacy of the sameQuantDiffUnit logic
Of course if the quantities don't match the two units aren't event the
same general quantities! I was thinking under the wrong assumption
that this would only be called when the quantities are definitely the
same. But we _also_ call it in `*` and `/` where the quantities can be
completely different!