do not emit `short` types in `to`, workaround weird issue
Emitting the short name is problematic for `min`, as it collides with
the `min` procedures and for some reason the nim compiler isn't smart
enough to figure out to use the type...

While doing that this leads to a bizarre bug, where the UnitTable is
somehow corrupted? Retrieving elements sometimes suddenly leads to
empty string data despite never including such a thing in the
table. If we detect that for now just delete that element and continue
constructing the identifier as normal.